add hosted graphite [#94181106]
