+++ .ruby-version
@@ -1,1 +1,1 @@
-rbx-2.3.0
+ruby-2.2.1
